passport_ocr/requirements.txt
pytesseract
opencv-python
Pillow
numpy
scikit-image